:bug: | Fix strange server in Menhera

+++ src/events/guildDelete.js
@@ -4,6 +4,7 @@ module.exports = class GuildDelete {
   }
 
   async run(guild) {
+    if (!guild || !guild.id) return;
     this.client.database.Guilds.findOneAndDelete({ id: guild.id }, () => {
       console.log(`[EVENT] Deleted Guild: ${guild.id}`);
     });
